frontend/python-attendance-checking-system/requirements.txt
customtkinter==5.1.2
Pillow==9.4.0
pycryptodome==3.17
pyscard==2.0.3
requests==2.28.2
smartcard==0.3
